+++ src/renderer/src/global.d.ts
@@ -9,6 +9,7 @@ export interface ResourceItem {
     | 'settings'
     | 'hook'
     | 'memory'
+    | 'file'
   name: string
   path: string | null
   description?: string